Button Class
MAIN UPDATE
Button Class and derivative classes

Works pretty much the same as the Background class. It has ButtonInput and ButtonAction classes that, again, work pretty much the same.

Two key differences:
ButtonAction's constructor takes an AssetType parameter which changes the button's action depending on the tag.
ButtonInput doesn't have a processKeyboardInput() function. It instead has a processMouseInput() function which takes mouse click inputs because its a button.

The Button class itself also takes an AssetType parameter which is used to set its sprite position on the screen, as well as to send to the ButtonAction class.

MISC UPDATES
- Added more AssetType tags to the enum class (MAIN_MENU_PLAY_BUTTON, MAIN_MENU_LEADERBOARD_BUTTON, MAIN_MENU_EXIT_BUTTON).
- Added createButtons() function to MainMenuScene class which creates 3 instances of the Button class.
- Added new texture loading to TextureManager class.
- Updated sprite for Main Menu background, added sprites for Main Menu buttons.
- Added Upward and Downward movement to Player class (not useful as of yet but it pissed me off so added it anyway).